requirements.txt
# Core Dependencies
PyQt5==5.15.7
Pillow==9.5.0
numpy==1.24.3

# Development Dependencies
pytest==7.3.1
pylint==2.17.4
black==23.3.0

# Optional: For future CLI support
click==8.1.3
